Create new test file for operation
